separate flask entrypoint from app definition
