+++ static.micahnorwoodjordan.com/startup/__init__.py
@@ -1,0 +1,11 @@
+from flask import Flask
+
+
+def create_app():
+    app = Flask(__name__)
+
+    # Register blueprints or import routes
+    from .routes import main as main_blueprint
+    app.register_blueprint(main_blueprint)
+
+    return app
